made snake move slower

+++ Snakes/Entity.java
@@ -1,9 +1,9 @@
 import java.awt.Graphics;
 
 public abstract class Entity{
-    protected float x,y;
+    protected double x,y;
 
-    public Entity(float x, float y){
+    public Entity(double x, double y){
       this.x = x;
       this.y = y;
     }
